requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confusable-homoglyphs==3.2.0
Django==4.1.7
django-crispy-forms-registration==0.1.3
django-dump-load-utf8==0.0.4
django-filter==22.1
django-registration==3.3
django-simple-history==3.3.0
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
